Park the SolverManager POC in the spring boot example until we are ready to write the API in stone
