+++ optaplanner-integration/optaplanner-spring-boot-example/src/main/java/org/optaplanner/spring/boot/example/poc/PocSolverManagerProducer.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2019 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.optaplanner.spring.boot.example.poc;
+
+import org.optaplanner.core.api.solver.SolverFactory;
+import org.optaplanner.spring.boot.example.poc.api.solver.SolverManager;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+// THIS IS JUST A PROOF OF CONCEPT IN THE EXAMPLE FOR 7.30.0.Final. THIS CLASS WILL BECOME OBSOLETE VERY SOON.
+// TODO Move this functionality into optaplanner-spring-boot-autoconfigure
+@Configuration
+public class PocSolverManagerProducer {
+
+    // TODO Move this method into OptaPlannerAutoConfiguration
+    @Bean
+    @ConditionalOnMissingBean
+    public <Solution_> SolverManager<Solution_> solverManager(SolverFactory<Solution_> solverFactory) {
+        // TODO supply ThreadFactory
+        return SolverManager.create(solverFactory);
+    }
+
+}
